added product controller and fixed delete feature

+++ backend/ProductManagementMicroservice/ProductManagement.Application/Features/ProductFeatures/DeleteProduct/DeleteProductHandler.cs
@@ -13,10 +13,14 @@ namespace ProductManagement.Application.Features.ProductFeatures.DeleteProduct
         }
         public async Task<Unit> Handle(DeleteProductRequest request, CancellationToken cancellationToken)
         {
-            var product = await _productRepository.GetByIdAsync(request.id);
+            var product = await _productRepository.GetByIdAsync(request.productId);
             if (product is null) 
             {
                 throw new NotFoundException("Product not found");
+            }
+            if(!product.CreatedUserId.Equals(request.userId)) 
+            {
+                throw new InvalidOperationException("This user can't delete this product");
             }
             await _productRepository.DeleteAsync(product);
             return Unit.Value;
